refactor: SidebarBackButton own folder
