+++ wondrous-app/components/Common/Sidebar/Entity/Role.tsx
@@ -1,5 +1,5 @@
 import { useQuery } from '@apollo/client';
-import BackButton from 'components/Common/Sidebar/Common/BackButton';
+import BackButton from 'components/Common/SidebarBackButton';
 import { useCanManage } from 'components/Common/Sidebar/Common/hooks';
 import Item from 'components/Common/Sidebar/Common/Item';
 import { AddIconWrapper, Label, ListWrapper } from 'components/Common/Sidebar/Common/styles';
